pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
  "pre-commit>=4.5.1",
  "pytest-repeat>=0.9.4",
  "pytest>=9.0.2",
  "pytest-asyncio>=1.3.0",
  "pytest-testmon>=2.2.0",
  "pytest-xdist>=3.8.0",
  "ruff>=0.14.14",
  "ty>=0.0.13",
  "testcontainers>=4.14.0",
  "pytest-playwright>=0.7.2",
  "toml-sort>=0.24.3"
]

[project]
name = "synthia"
version = "0.1.0"
description = "FastAPI application with Claude Agent SDK integration"
requires-python = ">=3.13"
dependencies = [
  "python-dotenv>=1.2.1",
  "fastapi[standard]>=0.128.0",
  "loguru>=0.7.3",
  "markdown>=3.10.1",
  "bleach>=6.3.0",
  "claude-agent-sdk==0.1.33",
  "mem0ai>=1.0.2",
  "psycopg[binary]>=3.3.2",
  "psycopg-pool>=3.3.0",
  "pydantic-settings>=2.12.0",
  "apscheduler>=3.11.2",
  "beautifulsoup4>=4.14.3",
  "table2ascii>=1.1.3",
  "google-genai>=1.60.0",
  "prometheus-fastapi-instrumentator>=7.1.0",
  "psutil>=7.2.1",
  "opentelemetry-api>=1.39.1",
  "opentelemetry-sdk>=1.39.1",
  "opentelemetry-exporter-otlp>=1.39.1",
  "opentelemetry-instrumentation-fastapi>=0.60b1",
  "opentelemetry-instrumentation-logging>=0.60b1",
  "asyncpg>=0.31.0",
  "numpy>=2.4.1",
  "anthropic>=0.76.0",
  "sentence-transformers>=5.2.0",
  "alembic>=1.18.1",
  "sqlalchemy>=2.0.46",
  "langchain-core>=1.2.7",
  "langchain-openai>=1.1.7",
  "pydantic>=2.12.5",
  "langsmith[claude-agent-sdk]>=0.6.4",
  "pytest-repeat>=0.9.4",
  "pywebpush>=2.2.0",
  "ollama>=0.6.1"
]

[tool.hatch.build.targets.wheel]
packages = ["synthia"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
  "smoke: marks tests as smoke tests (quick tests for basic functionality)",
  "voice: marks tests for voice/PWA functionality (requires running server)",
  "performance: marks performance benchmarks (excluded from make test)",
  "eval: marks evaluation tests for LLM capabilities"
]
filterwarnings = [
  "ignore::DeprecationWarning:testcontainers.*"
]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
select = [
  "E", # pycodestyle errors
  "W", # pycodestyle warnings
  "F", # pyflakes
  "I", # isort
  "B", # flake8-bugbear
  "C4", # flake8-comprehensions
  "UP" # pyupgrade
]

[tool.ruff.lint.isort]
known-first-party = ["synthia"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"synthia/main.py" = ["E402"]
